# xs2a/xs2a-api/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Reference Implementation of XS2A REST Services",
    "version" : "0.9.2-SNAPSHOT",
    "title" : "XS2A Services",
    "contact" : {
      "name" : "Andreas Reetz",
      "email" : "andreas.reetz@sdv-it.de"
    }
  },
  "host" : "api.sdv-it.de",
  "basePath" : "/tpp",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/{version}/accounts" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Read Account List (page 59) - reads all accounts belonging to PSU identified by OAuth2 Access Token",
        "description" : "",
        "operationId" : "readAccountList",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Consent-ID",
          "in" : "header",
          "description" : "Shall be contained if Consent Trasaction was performed via this API before",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, accounts returned",
            "schema" : {
              "$ref" : "#/definitions/AccountResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/accounts/{account-id}/balances" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Read Balance (page 61) - reads all balances according to specified account",
        "description" : "",
        "operationId" : "readBalance",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "account-id",
          "in" : "path",
          "description" : "Account ID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Consent-ID",
          "in" : "header",
          "description" : "Consent-ID as from Trasaction performed via this API before",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, balances returned",
            "schema" : {
              "$ref" : "#/definitions/AccountBalanceResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/accounts/{account-id}/transactions" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Read Transaction List (page 64) - reads all transactions according to specified account",
        "description" : "",
        "operationId" : "readTransactions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "account-id",
          "in" : "path",
          "description" : "Account ID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Consent-ID",
          "in" : "header",
          "description" : "Consent-ID as from Trasaction performed via this API before",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "date_from",
          "in" : "query",
          "description" : "Date from",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "date_to",
          "in" : "query",
          "description" : "Date to",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, transactions returned",
            "schema" : {
              "$ref" : "#/definitions/AccountTransactionsResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
    "/{version}/accounts/{account-id}/transactions/{page}" : {
      "get" : {
        "tags" : [ "Accounts" ],
        "summary" : "Read Transaction List (page 64) - reads all transactions according to specified account with pagination",
        "description" : "",
        "operationId" : "readTransactionsPage",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "path",
          "description" : "Version",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "account-id",
          "in" : "path",
          "description" : "Account ID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "path",
          "description" : "-- / first /second / current / last",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Process-ID",
          "in" : "header",
          "description" : "ID of the transaction as determined by the initiating party (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Request-ID",
          "in" : "header",
          "description" : "ID of the request (UUID)",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Consent-ID",
          "in" : "header",
          "description" : "Consent-ID as from Trasaction performed via this API before",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "PSU-ID",
          "in" : "header",
          "description" : "To be used if no OAuth Pre-Step was performed",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Authorization",
          "in" : "header",
          "description" : "Access Token as obtained from Authorization Server",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "signature",
          "in" : "header",
          "description" : "A signature of the request by the TPP on application level",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "certificate",
          "in" : "header",
          "description" : "The certificate used for signing the request",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "Date",
          "in" : "header",
          "description" : "Standard https header element for Date and Time",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "date_from",
          "in" : "query",
          "description" : "Date from",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "date_to",
          "in" : "query",
          "description" : "Date to",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK, transactions returned",
            "schema" : {
              "$ref" : "#/definitions/AccountTransactionsResponse"
            }
          },
          "400" : {
            "description" : "(Bad Request) Validation error occured"
          },
          "401" : {
            "description" : "(Unauthorized) Access Token is not valid"
          }
        }
      }
    },
...
  },
  "definitions" : {
    "Account" : {
      "type" : "object",
      "required" : [ "currency", "id" ],
      "properties" : {
        "id" : {
          "type" : "string",
          "example" : "3dc3d5b3-7023-4848-9853-f5400a64e80f",
          "description" : "Id of account"
        },
        "iban" : {
          "type" : "string",
          "example" : "DE2310010010123456789",
          "description" : "IBAN"
        },
        "bban" : {
          "type" : "string",
          "description" : "id for payment account without IBAN"
        },
        "pan" : {
          "type" : "string",
          "example" : "1234567890-12345",
          "description" : "Primary Account Number of a card"
        },
        "msisdn" : {
          "type" : "string",
          "description" : "An alias to access a payment account via a registered mobile phone number"
        },
        "name" : {
          "type" : "string",
          "description" : "Name given by the bank or the PSU in online banking"
        },
        "account_type" : {
          "type" : "string",
          "example" : "US dollar account",
          "description" : "Type of account"
        },
        "bic" : {
          "type" : "string",
          "description" : "The BIC associated to the account"
        },
        "balances" : {
          "type" : "string",
          "description" : "Balances"
        },
        "currency" : {
          "type" : "string",
          "example" : "EUR",
          "description" : "account currency"
        },
        "_links" : {
          "example" : "/v1/accounts/3dc3d5b3-7023-4848-9853-f5400a64e81g/balances",
          "description" : "Links to balances or transactions",
          "$ref" : "#/definitions/Links"
        }
      },
      "description" : "Account details"
    },
    "AccountAccount" : {
      "type" : "object",
      "properties" : {
        "iban" : {
          "type" : "string",
          "example" : "DE2310010010123456789",
          "description" : "IBAN"
        },
        "bban" : {
          "type" : "string",
          "description" : "id for payment account without IBAN"
        },
        "pan" : {
          "type" : "string",
          "example" : "1234567890-12345",
          "description" : "Primary Account Number of a card"
        },
        "msisdn" : {
          "type" : "string",
          "description" : "An alias to access a payment account via a registered mobile phone number"
        }
      }
    },
    "AccountAmount" : {
      "type" : "object",
      "required" : [ "content", "currency" ],
      "properties" : {
        "currency" : {
          "type" : "string",
          "example" : "EUR",
          "description" : "Abbreviation for currency"
        },
        "content" : {
          "type" : "string",
          "example" : "-100,00",
          "description" : "Amount"
        }
      }
    },
    "AccountBalance" : {
      "type" : "object",
      "properties" : {
        "booked" : {
          "description" : "Last known book balance of the account",
          "$ref" : "#/definitions/AccountBalanceInf"
        },
        "expected" : {
          "description" : "Balanced composed of booked entries and pending items",
          "$ref" : "#/definitions/AccountBalanceInf"
        },
        "authorised" : {
          "description" : "The expected balance together with the value of a pre-approved permanently available credit line",
          "$ref" : "#/definitions/AccountBalanceInf"
        },
        "opening_booked" : {
          "description" : "Balance at the beginning of the reporting period",
          "$ref" : "#/definitions/AccountBalanceInf"
        },
        "closing_booked" : {
          "description" : "Balance at the end of the pre-agreed reporting period",
          "$ref" : "#/definitions/AccountBalanceInf"
        },
        "interim_available" : {
          "description" : "Calculated on the basis of booked credit and debit items",
          "$ref" : "#/definitions/AccountBalanceInf"
        }
      }
    },
    "AccountBalanceInf" : {
      "type" : "object",
      "required" : [ "amount" ],
      "properties" : {
        "amount" : {
          "description" : "Amount and currency for balance for",
          "$ref" : "#/definitions/AccountAmount"
        },
        "last_action_date_time" : {
          "type" : "string"
        },
        "date" : {
          "type" : "string",
          "example" : "2017-11-01",
          "description" : "Date for last actualization"
        }
      }
    },
    "AccountBalanceResponse" : {
      "type" : "object",
      "properties" : {
        "balances" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/AccountBalance"
          }
        }
      },
      "description" : "Response for /{version}/accounts/{account-id}/balances request"
    },
...
    "AccountResponse" : {
      "type" : "object",
      "required" : [ "accounts" ],
      "properties" : {
        "accounts" : {
          "type" : "array",
          "description" : "List of accounts",
          "items" : {
            "$ref" : "#/definitions/Account"
          }
        }
      },
      "description" : "Response for /{version}/accounts request"
    },
    "AccountTransaction" : {
      "type" : "object",
      "required" : [ "amount", "booking_date", "remittance_information_unstructured", "transaction_id", "value_date" ],
      "properties" : {
        "transaction_id" : {
          "type" : "string",
          "example" : "1234567",
          "description" : "ID for transaction"
        },
        "creditor_name" : {
          "type" : "string",
          "example" : "John Miles",
          "description" : "Creditor name"
        },
        "creditor_account" : {
          "description" : "Creditor account",
          "$ref" : "#/definitions/AccountAccount"
        },
        "debtor_name" : {
          "type" : "string",
          "example" : "Paul Simpson",
          "description" : "Debtor name"
        },
        "debtor_account" : {
          "description" : "Debtor account",
          "$ref" : "#/definitions/AccountAccount"
        },
        "amount" : {
          "description" : "Amount",
          "$ref" : "#/definitions/AccountAmount"
        },
        "booking_date" : {
          "type" : "string",
          "example" : "2017-10-25",
          "description" : "Booking date"
        },
        "value_date" : {
          "type" : "string",
          "example" : "2017-10-26",
          "description" : "Value date"
        },
        "remittance_information_unstructured" : {
          "type" : "string",
          "example" : "Example for Remittance Information",
          "description" : "Remittance information"
        }
      }
    },
    "AccountTransactions" : {
      "type" : "object",
      "required" : [ "_links", "booked" ],
      "properties" : {
        "booked" : {
          "type" : "array",
          "description" : "transactions consider closed bookings",
          "items" : {
            "$ref" : "#/definitions/AccountTransaction"
          }
        },
        "pending" : {
          "type" : "array",
          "description" : "transactions consider planned bookings",
          "items" : {
            "$ref" : "#/definitions/AccountTransaction"
          }
        },
        "_links" : {
          "description" : "account_link / first_page_link / ... / last_page_link",
          "$ref" : "#/definitions/Links"
        }
      }
    },
    "AccountTransactionsResponse" : {
      "type" : "object",
      "properties" : {
        "transactions" : {
          "$ref" : "#/definitions/AccountTransactions"
        }
      },
      "description" : "Response for /{version}/accounts/{account-id}/transactions request"
    },
...
    "Links" : {
      "type" : "object",
      "properties" : {
        "self" : {
          "type" : "string",
          "description" : "Link to self"
        },
        "redirect" : {
          "type" : "string",
          "description" : "In case of an SCA Redirect Approach - the redirect link"
        },
        "update_psu_identification" : {
          "type" : "string",
          "description" : "Link to Update PSU Identification"
        },
        "update_psu_authentication" : {
          "type" : "string",
          "description" : "Link to Update PSU Authetication (mandatory if SCA approach is 'Embedded')"
        },
        "select_authentication_method" : {
          "type" : "string",
          "description" : "Link to resource where TPP can select authentication method"
        },
        "authorise_transaction" : {
          "type" : "string",
          "description" : "Link to resource where TPP can send 'Transaction Authorisation Request'"
        },
        "account_link" : {
          "type" : "string",
          "description" : "Link to an accounts/.../transactions"
        },
        "balances" : {
          "type" : "string",
          "example" : "/v1/accounts/3dc3d5b3-7023-4848-9853-f5400a64e80f/balances",
          "description" : "Link to balances of account is to return"
        },
        "transaction" : {
          "type" : "string",
          "example" : "/v1/accounts/3dc3d5b3-7023-4848-9853-f5400a64e80f/transactions",
          "description" : "Link to transactions of account is to return"
        },
        "status" : {
          "type" : "string",
          "description" : "Link to Status"
        },
        "first_page_link" : {
          "type" : "string",
          "description" : "Navigation link for accounts reports"
        },
        "second_page_link" : {
          "type" : "string",
          "description" : "Navigation link for accounts reports"
        },
        "current_page_link" : {
          "type" : "string",
          "description" : "Navigation link for accounts reports"
        },
        "last_page_link" : {
          "type" : "string",
          "description" : "Navigation link for accounts reports"
        }
      }
    },
...
  }
}